## Describe the changes this PR makes. Why should it be merged?

- Makes `Position` floating point (internally storing as radians `f64`)
rather than fixed-point, since we can no longer guarantee a tick LCM due
to supporting custom devices.
- Some minor additions to the `Position` type while we're at it
(`Position::ZERO`, `Position::from/as_gradians`).
- Adds a const-generic parameter to `AdiEncoder` for a custom encoder
resolution. Allows for easier use with custom encoder hardware.
- Adds the type alias `AdiOpticalEncoder` for creating `AdiEncoder<360>`
instances without specifying TPR explicitly (which most of the VRC kids
probably won't want to do).
- Changes `AdiGyro::yaw` to return heading in `f64` degrees rather than
a `Position`, since that was a mistake in the first place.

## Additional Context

Pretty breaking change.

+++ packages/vexide-devices/src/smart/motor.rs
@@ -323,7 +323,7 @@ impl Motor {
                 #[allow(clippy::cast_precision_loss)]
                 vexDeviceMotorAbsoluteTargetSet(
                     self.device,
-                    position.as_ticks(gearset.ticks_per_revolution()) as f64,
+                    position.as_ticks(gearset.ticks_per_revolution()),
                     velocity,
                 );
             },
@@ -812,7 +812,7 @@ impl Motor {
     pub fn position(&self) -> Result<Position, MotorError> {
         let gearset = self.gearset()?;
         Ok(Position::from_ticks(
-            unsafe { vexDeviceMotorPositionGet(self.device) } as i64,
+            unsafe { vexDeviceMotorPositionGet(self.device) },
             gearset.ticks_per_revolution(),
         ))
     }
@@ -969,7 +969,6 @@ impl Motor {
     ///     motor.set_position(Position::from_degrees(0.0)).unwrap();
     /// }
     /// ```
-    #[allow(clippy::cast_precision_loss)]
     pub fn set_position(&mut self, position: Position) -> Result<(), MotorError> {
         let gearset = self.gearset()?;
 
@@ -977,7 +976,7 @@ impl Motor {
             vexDeviceMotorPositionSet(
                 self.device,
                 // NOTE: No precision loss since ticks are not fractional.
-                position.as_ticks(gearset.ticks_per_revolution()) as f64,
+                position.as_ticks(gearset.ticks_per_revolution()),
             );
         }
